if block added to get site, need script for latest version

+++ init/60_get_the_site.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
-
+if [ ! -f "/config/www/owncloud/index.php" ]; then
 wget -O /tmp/owncloud.tar.bz2 https://download.owncloud.org/community/owncloud-8.1.1.tar.bz2
 cd /tmp
 tar -xjf owncloud.tar.bz2
 mv owncloud /config/www/
+rm -rf /tmp/*.bz2
+chown -R abc:abc /config/www/owncloud
+fi
